+++ smarc_mission_sim/package.xml
@@ -8,6 +8,7 @@
   <!-- Example:  -->
   <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
   <maintainer email="nilsbore@gmail.com">nbore</maintainer>
+  <maintainer email="ozkahramanozer@gmail.com">nbore</maintainer>
 
 
   <!-- One license tag required, multiple allowed, one license per tag -->
